pyproject.toml
[project]
name = "pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "distilabel[ollama]>=1.5.3",
    "openai>=1.107.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "tiktoken>=0.11.0",
    "torch>=2.8.0",
]
